Fully Implemented Forward & Backward Propagation
This version contains a fully functional serriform algorithm, which may be connected to other types of neural network layers to form a full network. More robust unit testing is needed in order to find bugs, and prove consistency. A mathematical proof is needed to officially derive the partial derivatives used in backpropagation.